% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_movies.R
\name{get_csv_movies_file}
\alias{get_csv_movies_file}
\title{Create a local copy of movies data}
\usage{
get_csv_movies_file()
}
\value{
logical TRUE if the operation was successful
}
\description{
This function will attempt to create a
local copy of the movies data set
such that you can make sure that you are
happy with reading in data from csv files.
}
