% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\docType{data}
\name{get_data}
\alias{get_data}
\title{Dummy data}
\usage{
data(get_data, package = "jrIntroduction")
}
\value{
A tibble
}
\description{
Dummy data set using in assignment 1
}
\examples{
data(get_data)
}
\keyword{datasets}
