% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExcelMovies.R
\name{GetExcelMovies}
\alias{GetExcelMovies}
\title{Create a local copy of movies data}
\usage{
GetExcelMovies()
}
\value{
logical TRUE if the operation was successful
}
\description{
This function will attempt to create a
local copy of the movies data set
such that you can make sure that you are
happy with reading in data from Excel files.
}
