% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{data}
\name{names}
\alias{names}
\title{Names data}
\usage{
data(names)
}
\value{
A named vector
}
\description{
Number of babies with a given name born in the US for year 2000.
}
\examples{
data(names)
}
\keyword{datasets}
