% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_names.R
\docType{data}
\name{USnames}
\alias{USnames}
\title{US baby names}
\usage{
data(USnames)
}
\value{
A data frame
}
\description{
A collection of names given to children born
 in the US during the years of 2011-2014
}
\examples{
data(walmartraw)
}
\keyword{datasets}
