% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walmartraw.R
\docType{data}
\name{walmartraw}
\alias{walmartraw}
\title{Walmart raw data}
\usage{
data(walmartraw)
}
\value{
A data frame
}
\description{
Used only for the graphic example on the slides
}
\examples{
data(walmartraw)
}
\keyword{datasets}
