% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{GetNumericVector}
\alias{GetNumericVector}
\alias{GetLogicalVector}
\alias{GetCharacterVector}
\title{Generate random vectors}
\usage{
GetNumericVector()

GetLogicalVector()

GetCharacterVector()
}
\description{
These functions create vectors of numerics, logicals or characters.
Given a seed, the functions will always return the
same subset.
}
